+++ .vscode/settings.json
@@ -1,6 +1,8 @@
 {
     "java.project.sourcePaths": [
+        "ParsingStrings/ParsingStrings/src",
         "OnlineShoppingCart/src",
-        "ParsingStrings/ParsingStrings/src"
+        "DateParser/DateParser/src",
+        "NameChanger/NameChanger/src"
     ]
 }
